backend/requirements.txt
Django~=3.1.7
djangorestframework
Pillow~=8.2.0
django-cors-headers